Lambda function code updated and renamed
